+++ src/main/java/ru/algotraide/component/FakeBalance.java
@@ -2,13 +2,14 @@ package ru.algotraide.component;
 
 import com.binance.api.client.domain.market.TickerPrice;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface FakeBalance {
-    Double getBalanceBySymbol(String symbol);
-    void setBalanceBySymbol(String symbol, Double value);
-    void addBalanceBySymbol(String symbol, Double value);
-    void reduceBalanceBySymbol(String symbol, Double value);
-    Double getAllBalanceInDollars(List<TickerPrice> prices);
+    BigDecimal getBalanceBySymbol(String symbol);
+    void setBalanceBySymbol(String symbol, BigDecimal value);
+    void addBalanceBySymbol(String symbol, BigDecimal value);
+    void reduceBalanceBySymbol(String symbol, BigDecimal value);
+    BigDecimal getAllBalanceInDollars(List<TickerPrice> prices);
     void resetBalance();
 }
